get_thread_dynamic_buffer(), get_thread_fixed_buffer(): widen size parm to guint
As guint16, a computed required size greater than 65535 silently wrapped,
returning an undersized buffer.

+++ src/util/glib_util.h
@@ -38,12 +38,12 @@ void *
 get_thread_dynamic_buffer(
       GPrivate * buf_key_ptr,
       GPrivate * bufsz_key_ptr,
-      guint16    required_size);
+      guint      required_size);
 
 void *
 get_thread_fixed_buffer(
       GPrivate * buf_key_ptr,
-      guint16    required_size);
+      guint      required_size);
 
 GPtrArray *
 gaux_ptr_array_truncate(
